refactor header tab
