+++ apps/dapp-console/app/components/Header/HeaderTab.tsx
@@ -1,5 +1,29 @@
+import { Route, routes } from '@/app/constants'
 import { cn } from '@/app/lib/utils'
 import Link from 'next/link'
+
+type HeaderTabsProps = {
+  currentRoute: Route['path']
+}
+
+const HeaderTabs = ({ currentRoute }: HeaderTabsProps) => {
+  return (
+    <div className="flex gap-8 items-end h-full">
+      {Object.keys(routes).map((route) => {
+        const { path, label } = routes[route]
+        return (
+          <HeaderTabItem
+            key={path}
+            href={path}
+            isActive={currentRoute === path}
+          >
+            {label}
+          </HeaderTabItem>
+        )
+      })}
+    </div>
+  )
+}
 
 type HeaderTabItemProps = {
   href: string
@@ -23,4 +47,4 @@ const HeaderTabItem = ({ href, isActive, children }: HeaderTabItemProps) => {
   )
 }
 
-export { HeaderTabItem }
+export { HeaderTabs, HeaderTabItem }
