+++ apps/dapp-console/app/components/Header/index.tsx
@@ -4,8 +4,7 @@ import Image from 'next/image'
 import { ThemeToggle } from '@/app/components/Header/ThemeToggle'
 import { Separator } from '@eth-optimism/ui-components/src/components/ui/separator'
 import { Button } from '@eth-optimism/ui-components/src/components/ui/button'
-import { HeaderTabItem } from './HeaderTab'
-import { Route, routes } from '@/app/constants'
+import { HeaderTabs } from './HeaderTab'
 import { usePathname } from 'next/navigation'
 
 const Header = () => {
@@ -37,29 +36,6 @@ const HeaderLogo = () => {
       />
       <Separator orientation="vertical" className="h-4 mx-4" />
       <span className="tracking-widest font-medium">DAPP DEVELOPER</span>
-    </div>
-  )
-}
-
-type HeaderTabsProps = {
-  currentRoute: Route['path']
-}
-
-const HeaderTabs = ({ currentRoute }: HeaderTabsProps) => {
-  return (
-    <div className="flex gap-8 items-end h-full">
-      {Object.keys(routes).map((route) => {
-        const { path, label } = routes[route]
-        return (
-          <HeaderTabItem
-            key={path}
-            href={path}
-            isActive={currentRoute === path}
-          >
-            {label}
-          </HeaderTabItem>
-        )
-      })}
     </div>
   )
 }
